+++ R/globals.R
@@ -6,7 +6,9 @@ utils::globalVariables(c("qol_ID", "qol_from", "qol_to", "from", "to", "delta",
                          "across", ":=", "TYPE", "TYPE_NR", "DEPTH", ".subheader",
                          "var_sum", "var_cum_sum", "pct_group", "var_cum_pct",
                          "mean", "sd", "min", "max", "sum_wgt", ".temp_key", ".temp_weight",
-                         "var_pct_row", ".pseudo_preserve", "BY", "by_vars", "VALUE"))
+                         "var_pct_row", ".pseudo_preserve", "BY", "by_vars", "VALUE",
+                         "first_person", "age", "age_factor", "income", "income_factor",
+                         "expenses", "expenses_factor"))
 
 
 # Start up message