+++ tests/testthat/test-any_table.R
@@ -40,7 +40,7 @@ test_that("Simplest form of any_table", {
     expect_equal(length(result_list), 3)
     expect_equal(names(result_list[[1]]), c("row.label", "var1", "weight_sum_1",
                                             "weight_sum_2", "weight_sum_NA"))
-    expect_equal(result_list[[1]][["var1"]], c(paste0(0:99), "."))
+    expect_equal(result_list[[1]][["var1"]][1:90], as.character(0:89))
 })
 
 
@@ -62,6 +62,7 @@ test_that("any_table with multiple combinations", {
           any_table(rows    = c("age", "age + education"),
                     columns = c("sex + year", "sex"),
                     values  = weight,
+                    output  = "excel_nostyle",
                     print   = FALSE))
 
     expect_type(result_list, "list")
@@ -75,6 +76,7 @@ test_that("any_table many combinations don't break", {
                                   "state + age", "education + age"),
                       columns = c("year", "sex + year", "sex"),
                       values  = weight,
+                      output  = "excel_nostyle",
                       print   = FALSE))
 
     expect_type(result_list, "list")
@@ -176,6 +178,7 @@ test_that("any_table with interleaved order", {
                       values     = weight,
                       statistics = c("sum", "freq", "missing"),
                       order_by   = "interleaved",
+                      output     = "excel_nostyle",
                       print      = FALSE))
 
     expect_equal(names(result_list[[1]]),
@@ -186,38 +189,49 @@ test_that("any_table with interleaved order", {
 
 
 test_that("any_table with values order", {
+    year_current <- as.integer(format(Sys.Date(), "%Y"))
+
     result_list <- suppressMessages(dummy_df |>
+            collapse::fsubset(year < year_current) |>
             any_table(rows       = "age",
                       columns    = c("sex", "year"),
                       values     = c(weight, income),
                       statistics = "sum",
                       order_by   = "values",
+                      output     = "excel_nostyle",
                       print      = FALSE))
-
-    year <- as.integer(format(Sys.Date(), "%Y"))
 
     expect_equal(names(result_list[[1]]),
                  c("row.label", "var1", "weight_sum_1", "weight_sum_2", "weight_sum_NA",
-                   paste0("weight_sum_", year - 2), paste0("weight_sum_", year - 1), "income_sum_1", "income_sum_2",
-                   "income_sum_NA", paste0("income_sum_", year - 2), paste0("income_sum_", year - 1)))
+                   paste0("weight_sum_", year_current - 4), paste0("weight_sum_", year_current - 3),
+                   paste0("weight_sum_", year_current - 2), paste0("weight_sum_", year_current - 1),
+                   "income_sum_1", "income_sum_2", "income_sum_NA", paste0("income_sum_", year_current - 4),
+                   paste0("income_sum_", year_current - 3), paste0("income_sum_", year_current - 2),
+                   paste0("income_sum_", year_current - 1)))
 })
 
 
 test_that("any_table with columns order", {
+    year_current <- as.integer(format(Sys.Date(), "%Y"))
+
     result_list <- suppressMessages(dummy_df |>
+            collapse::fsubset(year < year_current) |>
             any_table(rows       = "age",
                       columns    = c("sex", "year"),
                       values     = c(weight, income),
                       statistics = "sum",
                       order_by   = "columns",
+                      output     = "excel_nostyle",
                       print      = FALSE))
-
-    year <- as.integer(format(Sys.Date(), "%Y"))
 
     expect_equal(names(result_list[[1]]),
                  c("row.label", "var1", "weight_sum_1", "weight_sum_2", "weight_sum_NA",
-                   "income_sum_1", "income_sum_2", "income_sum_NA", paste0("weight_sum_", year - 2),
-                   paste0("weight_sum_", year - 1), paste0("income_sum_", year - 2), paste0("income_sum_", year - 1)))
+                   "income_sum_1", "income_sum_2", "income_sum_NA",
+                   paste0("weight_sum_", year_current - 4), paste0("weight_sum_", year_current - 3),
+                   paste0("weight_sum_", year_current - 2), paste0("weight_sum_", year_current - 1),
+                   paste0("income_sum_", year_current - 4),
+                   paste0("income_sum_", year_current - 3), paste0("income_sum_", year_current - 2),
+                   paste0("income_sum_", year_current - 1)))
 })
 
 
@@ -254,6 +268,7 @@ test_that("any_table with by variables and multiple row and column variables", {
                      values  = weight,
                      by      = year,
                      print   = FALSE,
+                     output  = "excel_nostyle",
                      na.rm   = TRUE)
 
     expect_true("BY" %in% names(result_list[["table"]]))