+++ tests/testthat/test-variable_selection.R
@@ -18,23 +18,23 @@ test_that("Get inverse variable names from data frame works correct", {
 
 
 test_that("Get variable names between two variables", {
-    var_names <- dummy_df |> vars_between(state, income)
+    var_names <- dummy_df |> vars_between(state, person_id)
 
     expect_equal(length(var_names), 5)
 })
 
 
 test_that("Get variable names from variable to the end", {
-    var_names <- dummy_df |> vars_between(state)
+    var_names <- dummy_df |> vars_between(body_weight)
 
     expect_equal(length(var_names), 7)
 })
 
 
 test_that("Get variable names from beginning to position", {
-    var_names <- dummy_df |> vars_between(to = state)
+    var_names <- dummy_df |> vars_between(to = person_id)
 
-    expect_equal(length(var_names), 5)
+    expect_equal(length(var_names), 6)
 })
 
 ###############################################################################
@@ -42,22 +42,22 @@ test_that("Get variable names from beginning to position", {
 ###############################################################################
 
 test_that("Providing not existing from-variable throws a warning", {
-    expect_message(var_names <- dummy_df |> vars_between(test, state), " ! WARNING: 'from' variable '")
+    expect_message(var_names <- dummy_df |> vars_between(test, person_id), " ! WARNING: 'from' variable '")
 
-    expect_equal(length(var_names), 5)
+    expect_equal(length(var_names), 6)
 })
 
 
 test_that("Providing not existing to-variable throws a warning", {
     expect_message(var_names <- dummy_df |> vars_between(state, test), " ! WARNING: 'to' variable '")
 
-    expect_equal(length(var_names), 7)
+    expect_equal(length(var_names), 18)
 })
 
 
-test_that("Providing more than one from or to variable throws a ntoe", {
+test_that("Providing more than one from or to variable throws a note", {
     expect_message(var_names <- dummy_df |> vars_between(c(state, age), income),
                    " ~ NOTE: Only single variable names allowed for 'from' and 'to'. The respective\n")
 
-    expect_equal(length(var_names), 5)
+    expect_equal(length(var_names), 14)
 })